% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spellcheckr.R
\docType{data}
\name{dict}
\alias{dict}
\alias{splits}
\title{dict}
\format{A data frame with 245462 rows and 2 columns}
\usage{
data(dict)
}
\description{
A dataset containing the valid english dictionary words originally obtained from a work of Jane Austen.
}
\details{
\itemize{
  \item . The actual word
  \item N The number of occurences of the word in the document.
}
}
\keyword{datasets}

